+++ src/resource.h
@@ -201,8 +201,9 @@ protected:
 template <typename T>
 class PreemptiveResource: public Resource {
 public:
-  PreemptiveResource(Simulator* sim, std::string name, int mon, int capacity, int queue_size):
-    Resource(sim, name, mon, capacity, queue_size) {}
+  PreemptiveResource(Simulator* sim, std::string name, int mon, 
+                     int capacity, int queue_size, bool keep_queue):
+    Resource(sim, name, mon, capacity, queue_size), keep_queue(keep_queue) {}
   
   virtual void reset() {
     Resource::reset();
@@ -213,6 +214,7 @@ public:
   }
   
 protected:
+  bool keep_queue;
   RPQueue preempted;    /**< preempted arrivals */
   T server;             /**< server container */
   
@@ -237,8 +239,16 @@ protected:
       double last = first->arrival->get_activity(this->name);
       first->arrival->set_activity(this->name, time - last);
     }
-    preempted.insert((*first));
-    queue_count += first->amount;
+    if (keep_queue) {
+      if (!room_in_queue(first->amount, first->priority)) {
+        if (verbose) verbose_print(time, first->arrival->name, "REJECT");
+        first->arrival->terminate(time, false);
+      } else insert_in_queue(verbose, time, first->arrival, first->amount, 
+                             first->priority, first->preemptible, first->restart);
+    } else {
+      preempted.insert((*first));
+      queue_count += first->amount;
+    }
     server_count -= first->amount;
     server.erase(first);
     return true;