+++ tests/testthat/test-simmer-resource-preemption.R
@@ -5,9 +5,8 @@ test_that("a lower priority arrival gets rejected before accessing the server", 
     seize("dummy", 1) %>%
     timeout(10) %>%
     release("dummy", 1)
-  
   t1 <- create_trajectory() %>%
-    seize("dummy", 1, priority=1, preemptible=1) %>%
+    seize("dummy", 1, priority=1) %>%
     timeout(10) %>%
     release("dummy", 1)
   
@@ -29,9 +28,8 @@ test_that("tasks are NOT restarted", {
     seize("dummy", 1, restart=FALSE) %>%
     timeout(10) %>%
     release("dummy", 1)
-  
   t1 <- create_trajectory() %>%
-    seize("dummy", 2, priority=1, preemptible=1) %>%
+    seize("dummy", 2, priority=1) %>%
     timeout(10) %>%
     release("dummy", 2)
   
@@ -54,9 +52,8 @@ test_that("tasks are restarted", {
     seize("dummy", 1, restart=TRUE) %>%
     timeout(10) %>%
     release("dummy", 1)
-  
   t1 <- create_trajectory() %>%
-    seize("dummy", 2, priority=1, preemptible=1) %>%
+    seize("dummy", 2, priority=1) %>%
     timeout(10) %>%
     release("dummy", 2)
   
@@ -78,9 +75,8 @@ test_that("tasks are preempted in a FIFO basis", {
     seize("dummy", 1, restart=TRUE) %>%
     timeout(10) %>%
     release("dummy", 1)
-  
   t1 <- create_trajectory() %>%
-    seize("dummy", 1, priority=1, preemptible=1) %>%
+    seize("dummy", 1, priority=1) %>%
     timeout(10) %>%
     release("dummy", 1)
   
@@ -102,9 +98,8 @@ test_that("tasks are preempted in a LIFO basis", {
     seize("dummy", 1, restart=TRUE) %>%
     timeout(10) %>%
     release("dummy", 1)
-  
   t1 <- create_trajectory() %>%
-    seize("dummy", 1, priority=1, preemptible=1) %>%
+    seize("dummy", 1, priority=1) %>%
     timeout(10) %>%
     release("dummy", 1)
   
@@ -119,4 +114,93 @@ test_that("tasks are preempted in a LIFO basis", {
   
   expect_equal(arrs_ordered$end_time, c(22, 23, 12, 13))
   expect_equal(arrs_ordered$activity_time, c(13, 11, 10, 10))
+})
+
+test_that("queue can exceed queue_size by default", {
+  t0 <- create_trajectory() %>%
+    seize("dummy", 1) %>%
+    timeout(10) %>%
+    release("dummy")
+  t1 <- create_trajectory() %>%
+    seize("dummy", priority=1) %>%
+    timeout(10) %>%
+    release("dummy")
+  
+  env <- simmer() %>%
+    add_generator("p0a", t0, at(0, 0)) %>%
+    add_generator("p1a", t1, at(1)) %>%
+    add_resource("dummy", 1, 1, preemptive=TRUE) %>%
+    run()
+  
+  res <- env %>% get_mon_resources()
+  arr <- env %>% get_mon_arrivals()
+  arr_ordered <- arr[order(arr$name),]
+  
+  expect_equal(res$time, c(0, 0, 1, 11, 20, 30))
+  expect_equal(res$server, c(1, 1, 1, 1, 1, 0))
+  expect_equal(res$queue, c(0, 1, 2, 1, 0, 0))
+  expect_equal(arr_ordered$end_time, c(20, 30, 11))
+  expect_equal(arr_ordered$activity_time, c(10, 10, 10))
+  expect_equal(arr_ordered$finished, c(TRUE, TRUE, TRUE))
+})
+
+test_that("queue cannot exceed queue_size with hard limit (preempted rejected)", {
+  t0 <- create_trajectory() %>%
+    seize("dummy", 1) %>%
+    timeout(10) %>%
+    release("dummy")
+  t1 <- create_trajectory() %>%
+    seize("dummy", priority=1) %>%
+    timeout(10) %>%
+    release("dummy")
+  
+  env <- simmer() %>%
+    add_generator("p0a", t0, at(0, 0)) %>%
+    add_generator("p1a", t1, at(1)) %>%
+    add_resource("dummy", 1, 1, preemptive=TRUE, queue_size_strict=TRUE) %>%
+    run()
+  
+  res <- env %>% get_mon_resources()
+  arr <- env %>% get_mon_arrivals()
+  arr_ordered <- arr[order(arr$name),]
+  
+  expect_equal(res$time, c(0, 0, 1, 11, 21))
+  expect_equal(res$server, c(1, 1, 1, 1, 0))
+  expect_equal(res$queue, c(0, 1, 1, 0, 0))
+  expect_equal(arr_ordered$end_time, c(1, 21, 11))
+  expect_equal(arr_ordered$activity_time, c(1, 10, 10))
+  expect_equal(arr_ordered$finished, c(FALSE, TRUE, TRUE))
+})
+
+test_that("queue cannot exceed queue_size with hard limit (preempted to queue)", {
+  t0 <- create_trajectory() %>%
+    seize("dummy", 1) %>%
+    timeout(10) %>%
+    release("dummy")
+  t1 <- create_trajectory() %>%
+    seize("dummy", priority=1) %>%
+    timeout(10) %>%
+    release("dummy")
+  t2 <- create_trajectory() %>%
+    seize("dummy", priority=2) %>%
+    timeout(10) %>%
+    release("dummy")
+  
+  env <- simmer() %>%
+    add_generator("p0a", t0, at(0)) %>%
+    add_generator("p1a", t1, at(0)) %>%
+    add_generator("p2a", t2, at(1)) %>%
+    add_resource("dummy", 1, 1, preemptive=TRUE, queue_size_strict=TRUE) %>%
+    run()
+  
+  res <- env %>% get_mon_resources()
+  arr <- env %>% get_mon_arrivals()
+  arr_ordered <- arr[order(arr$name),]
+  
+  expect_equal(res$time, c(0, 0, 1, 11, 20))
+  expect_equal(res$server, c(1, 1, 1, 1, 0))
+  expect_equal(res$queue, c(0, 1, 1, 0, 0))
+  expect_equal(arr_ordered$end_time, c(1, 20, 11))
+  expect_equal(arr_ordered$activity_time, c(0, 10, 10))
+  expect_equal(arr_ordered$finished, c(FALSE, TRUE, TRUE))
 })